Refactor channel module, validate FEM, add physical constants
