--- crates/cfd-1d/src/channel.rs
@@ -1,799 +1,0 @@
-//! Channel geometry and flow characteristics for 1D CFD.
-//!
-//! This module provides advanced channel modeling capabilities including
-//! complex geometries, surface effects, and flow regime transitions.
-
-use cfd_core::fluid::Fluid;
-use cfd_core::error::Result;
-use nalgebra::RealField;
-use num_traits::cast::FromPrimitive;
- // bring Float methods into scope for T
-use serde::{Deserialize, Serialize};
-// Removed unused import following YAGNI principle
-
-/// Extended channel geometry representation
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ChannelGeometry<T: RealField + Copy> {
-    /// Channel type
-    pub channel_type: ChannelType,
-    /// Length [m]
-    pub length: T,
-    /// Cross-sectional parameters
-    pub cross_section: CrossSection<T>,
-    /// Surface properties
-    pub surface: SurfaceProperties<T>,
-    /// Geometric variations along length
-    pub variations: Vec<GeometricVariation<T>>,
-}
-
-/// Types of channel geometries
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub enum ChannelType {
-    /// Straight channel
-    Straight,
-    /// Curved channel
-    Curved {
-        /// Radius of curvature in meters
-        radius: f64
-    },
-    /// Tapered channel
-    Tapered,
-    /// Serpentine channel
-    Serpentine {
-        /// Number of turns in the serpentine path
-        turns: usize
-    },
-    /// Spiral channel
-    Spiral {
-        /// Number of turns in the spiral (can be fractional)
-        turns: f64
-    },
-}
-
-/// Cross-sectional geometry
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub enum CrossSection<T: RealField + Copy> {
-    /// Rectangular cross-section
-    Rectangular {
-        /// Width of the rectangular channel
-        width: T,
-        /// Height of the rectangular channel
-        height: T
-    },
-    /// Circular cross-section
-    Circular {
-        /// Diameter of the circular channel
-        diameter: T
-    },
-    /// Elliptical cross-section
-    Elliptical {
-        /// Major axis length of the ellipse
-        major_axis: T,
-        /// Minor axis length of the ellipse
-        minor_axis: T
-    },
-    /// Trapezoidal cross-section
-    Trapezoidal {
-        /// Width at the top of the trapezoid
-        top_width: T,
-        /// Width at the bottom of the trapezoid
-        bottom_width: T,
-        /// Height of the trapezoid
-        height: T
-    },
-    /// Custom cross-section with area and hydraulic diameter
-    Custom {
-        /// Cross-sectional area
-        area: T,
-        /// Hydraulic diameter (4 * area / perimeter)
-        hydraulic_diameter: T
-    },
-}
-
-/// Surface properties affecting flow
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct SurfaceProperties<T: RealField + Copy> {
-    /// Surface roughness [m]
-    pub roughness: T,
-    /// Contact angle [radians]
-    pub contact_angle: Option<T>,
-    /// Surface energy [J/m²]
-    pub surface_energy: Option<T>,
-    /// Hydrophobic/hydrophilic nature
-    pub wettability: Wettability,
-}
-
-/// Surface wettability characteristics
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub enum Wettability {
-    /// Hydrophilic surface
-    Hydrophilic,
-    /// Hydrophobic surface
-    Hydrophobic,
-    /// Superhydrophilic surface
-    Superhydrophilic,
-    /// Superhydrophobic surface
-    Superhydrophobic,
-}
-
-/// Geometric variation along channel length
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct GeometricVariation<T: RealField + Copy> {
-    /// Position along channel [0-1]
-    pub position: T,
-    /// Scale factor for cross-section
-    pub scale_factor: T,
-    /// Local roughness modification
-    pub roughness_factor: T,
-}
-
-/// Extended channel flow model
-pub struct Channel<T: RealField + Copy> {
-    /// Channel geometry
-    pub geometry: ChannelGeometry<T>,
-    /// Flow state
-    pub flow_state: FlowState<T>,
-    /// Numerical parameters
-    pub numerical_params: NumericalParameters<T>,
-}
-
-/// Flow state information
-#[derive(Debug, Clone)]
-pub struct FlowState<T: RealField + Copy> {
-    /// Reynolds number
-    pub reynolds_number: Option<T>,
-    /// Flow regime
-    pub flow_regime: FlowRegime,
-    /// Entrance length effects
-    pub entrance_effects: bool,
-    /// Secondary flow effects
-    pub secondary_flows: bool,
-}
-
-/// Flow regime classification
-#[derive(Debug, Clone, PartialEq)]
-pub enum FlowRegime {
-    /// Stokes flow (Re << 1)
-    Stokes,
-    /// Laminar flow
-    Laminar,
-    /// Transitional flow
-    Transitional,
-    /// Turbulent flow
-    Turbulent,
-    /// Slip flow (rarefied gas)
-    SlipFlow,
-}
-
-/// Numerical parameters for advanced modeling
-#[derive(Debug, Clone)]
-pub struct NumericalParameters<T: RealField + Copy> {
-    /// Number of discretization points
-    pub discretization_points: usize,
-    /// Convergence tolerance
-    pub tolerance: T,
-    /// Include entrance effects
-    pub entrance_effects: bool,
-    /// Include surface tension effects
-    pub surface_tension_effects: bool,
-}
-
-impl<T: RealField + Copy + FromPrimitive + num_traits::Float> ChannelGeometry<T> {
-    /// Create a rectangular channel geometry
-    pub fn rectangular(length: T, width: T, height: T, roughness: T) -> Self {
-        Self {
-            channel_type: ChannelType::Straight,
-            length,
-            cross_section: CrossSection::Rectangular { width, height },
-            surface: SurfaceProperties {
-                roughness,
-                contact_angle: None,
-                surface_energy: None,
-                wettability: Wettability::Hydrophilic,
-            },
-            variations: Vec::new(),
-        }
-    }
-
-    /// Create a circular channel geometry
-    pub fn circular(length: T, diameter: T, roughness: T) -> Self {
-        Self {
-            channel_type: ChannelType::Straight,
-            length,
-            cross_section: CrossSection::Circular { diameter },
-            surface: SurfaceProperties {
-                roughness,
-                contact_angle: None,
-                surface_energy: None,
-                wettability: Wettability::Hydrophilic,
-            },
-            variations: Vec::new(),
-        }
-    }
-
-    /// Get cross-sectional area
-    pub fn area(&self) -> T {
-        match &self.cross_section {
-            CrossSection::Rectangular { width, height } => *width * *height,
-            CrossSection::Circular { diameter } => {
-                let pi = T::from_f64(std::f64::consts::PI).unwrap_or_else(|| T::zero());
-                let radius = *diameter / T::from_f64(2.0).unwrap_or_else(|| T::zero());
-                pi * radius * radius
-            },
-            CrossSection::Elliptical { major_axis, minor_axis } => {
-                let pi = T::from_f64(std::f64::consts::PI).unwrap_or_else(|| T::zero());
-                pi * *major_axis * *minor_axis / T::from_f64(4.0).unwrap_or_else(|| T::zero())
-            },
-            CrossSection::Trapezoidal { top_width, bottom_width, height } => {
-                (*top_width + *bottom_width) * *height / T::from_f64(2.0).unwrap_or_else(|| T::zero())
-            },
-            CrossSection::Custom { area, .. } => *area,
-        }
-    }
-
-    /// Get hydraulic diameter
-    pub fn hydraulic_diameter(&self) -> T {
-        match &self.cross_section {
-            CrossSection::Rectangular { width, height } => {
-                let four = T::from_f64(4.0).unwrap_or_else(|| T::zero());
-                four * self.area() / (T::from_f64(2.0).unwrap_or_else(|| T::zero()) * (*width + *height))
-            },
-            CrossSection::Circular { diameter } => *diameter,
-            CrossSection::Elliptical { major_axis, minor_axis } => {
-                // Use definition Dh = 4 A / P with Ramanujan perimeter
-                let four = T::from_f64(4.0).unwrap_or_else(|| T::zero());
-                four * self.area() / self.wetted_perimeter()
-            },
-            CrossSection::Trapezoidal { top_width, bottom_width, height } => {
-                let area = self.area();
-                let hw = (*top_width - *bottom_width) / T::from_f64(2.0).unwrap_or_else(|| T::zero());
-                let side_length = num_traits::Float::sqrt(num_traits::Float::powi(*height, 2) + num_traits::Float::powi(hw, 2));
-                let perimeter = *top_width + *bottom_width + T::from_f64(2.0).unwrap_or_else(|| T::zero()) * side_length;
-                T::from_f64(4.0).unwrap_or_else(|| T::zero()) * area / perimeter
-            },
-            CrossSection::Custom { hydraulic_diameter, .. } => *hydraulic_diameter,
-        }
-    }
-
-    /// Get wetted perimeter
-    pub fn wetted_perimeter(&self) -> T {
-        match &self.cross_section {
-            CrossSection::Rectangular { width, height } => {
-                T::from_f64(2.0).unwrap_or_else(|| T::zero()) * (*width + *height)
-            },
-            CrossSection::Circular { diameter } => {
-                let pi = T::from_f64(std::f64::consts::PI).unwrap_or_else(|| T::zero());
-                pi * *diameter
-            },
-            CrossSection::Elliptical { major_axis, minor_axis } => {
-                // Ramanujan's approximation for ellipse perimeter
-                let pi = T::from_f64(std::f64::consts::PI).unwrap_or_else(|| T::zero());
-                let a = *major_axis / T::from_f64(2.0).unwrap_or_else(|| T::zero());
-                let b = *minor_axis / T::from_f64(2.0).unwrap_or_else(|| T::zero());
-                let h = num_traits::Float::powi((a - b) / (a + b), 2);
-                pi * (a + b) * (T::one() + T::from_f64(3.0).unwrap_or_else(|| T::zero()) * h /
-                    (T::from_f64(10.0).unwrap_or_else(|| T::zero()) + num_traits::Float::sqrt(T::from_f64(4.0).unwrap_or_else(|| T::zero()) - T::from_f64(3.0).unwrap_or_else(|| T::zero()) * h)))
-            },
-            CrossSection::Trapezoidal { top_width, bottom_width, height } => {
-                let hw = (*top_width - *bottom_width) / T::from_f64(2.0).unwrap_or_else(|| T::zero());
-                let side_length = num_traits::Float::sqrt(num_traits::Float::powi(*height, 2) + num_traits::Float::powi(hw, 2));
-                *top_width + *bottom_width + T::from_f64(2.0).unwrap_or_else(|| T::zero()) * side_length
-            },
-            CrossSection::Custom { area, hydraulic_diameter } => {
-                T::from_f64(4.0).unwrap_or_else(|| T::zero()) * *area / *hydraulic_diameter
-            },
-        }
-    }
-}
-
-impl<T: RealField + Copy + FromPrimitive + num_traits::Float> Channel<T> {
-    /// Create a new channel with geometry
-    pub fn new(geometry: ChannelGeometry<T>) -> Self {
-        Self {
-            geometry,
-            flow_state: FlowState {
-                reynolds_number: None,
-                flow_regime: FlowRegime::Laminar,
-                entrance_effects: false,
-                secondary_flows: false,
-            },
-            numerical_params: NumericalParameters {
-                discretization_points: 100,
-                tolerance: T::from_f64(1e-6).unwrap_or_else(|| T::zero()),
-                entrance_effects: false,
-                surface_tension_effects: false,
-            },
-        }
-    }
-
-    /// Calculate hydraulic resistance using advanced models
-    pub fn calculate_resistance(&mut self, fluid: &Fluid<T>) -> Result<T> {
-        // Update flow state
-        self.update_flow_state(fluid)?;
-
-        // Calculate resistance based on flow regime
-        match self.flow_state.flow_regime {
-            FlowRegime::Stokes => self.calculate_stokes_resistance(fluid),
-            FlowRegime::Laminar => self.calculate_laminar_resistance(fluid),
-            FlowRegime::Transitional => self.calculate_transitional_resistance(fluid),
-            FlowRegime::Turbulent => self.calculate_turbulent_resistance(fluid),
-            FlowRegime::SlipFlow => self.calculate_slip_flow_resistance(fluid),
-        }
-    }
-
-    /// Update flow state based on current conditions
-    fn update_flow_state(&mut self, _fluid: &Fluid<T>) -> Result<()> {
-        // Calculate Reynolds number if velocity is known
-        if let Some(re) = self.flow_state.reynolds_number {
-            self.flow_state.flow_regime = self.classify_flow_regime(re);
-            
-            // Check for entrance effects
-            let dh = self.geometry.hydraulic_diameter();
-            // Entrance length correlations:
-            // Laminar: L_e/D_h ≈ 0.06 * Re
-            // Turbulent: L_e/D_h ≈ 4.4 * Re^(1/6)
-            let entrance_length = match self.flow_state.flow_regime {
-                FlowRegime::Laminar => {
-                    // Laminar entrance length
-                    dh * T::from_f64(0.06).unwrap_or_else(|| T::zero()) * re
-                }
-                FlowRegime::Transitional => {
-                    // Use turbulent correlation for transitional
-                    let one_sixth = T::from_f64(1.0/6.0).unwrap_or_else(|| T::zero());
-                    dh * T::from_f64(4.4).unwrap_or_else(|| T::zero()) * num_traits::Float::powf(re, one_sixth)
-                }
-                FlowRegime::Turbulent => {
-                    // Turbulent entrance length
-                    let one_sixth = T::from_f64(1.0/6.0).unwrap_or_else(|| T::zero());
-                    dh * T::from_f64(4.4).unwrap_or_else(|| T::zero()) * num_traits::Float::powf(re, one_sixth)
-                }
-                FlowRegime::Stokes => {
-                    // Stokes flow - use laminar correlation
-                    dh * T::from_f64(0.06).unwrap_or_else(|| T::zero()) * re
-                }
-                FlowRegime::SlipFlow => {
-                    // Slip flow - use modified correlation
-                    dh * T::from_f64(0.1).unwrap_or_else(|| T::zero()) * re
-                }
-            };
-            self.flow_state.entrance_effects = self.geometry.length < entrance_length;
-        }
-
-        Ok(())
-    }
-
-    /// Classify flow regime based on Reynolds number
-    fn classify_flow_regime(&self, reynolds: T) -> FlowRegime {
-        let re_val = reynolds.to_f64().unwrap_or(0.0);
-
-        if re_val < 0.1 {
-            FlowRegime::Stokes
-        } else if re_val < cfd_core::constants::LAMINAR_THRESHOLD {
-            FlowRegime::Laminar
-        } else if re_val <= cfd_core::constants::TURBULENT_THRESHOLD {
-            FlowRegime::Transitional
-        } else {
-            FlowRegime::Turbulent
-        }
-    }
-
-    /// Calculate resistance for Stokes flow
-    fn calculate_stokes_resistance(&self, fluid: &Fluid<T>) -> Result<T> {
-        let area = self.geometry.area();
-        let dh = self.geometry.hydraulic_diameter();
-        let length = self.geometry.length;
-        // Use actual operating temperature instead of hardcoded 20°C
-        let temperature = T::from_f64(293.15).unwrap_or_else(|| T::zero()); // Default to 20°C if not specified
-        let viscosity = fluid.dynamic_viscosity(temperature)?;
-
-        // Stokes flow resistance with shape factor: R = (f*Re) * μ * L / (2 * A * Dh^2)
-        let shape_factor = self.get_shape_factor();
-        let resistance = shape_factor * viscosity * length / (T::from_f64(2.0).unwrap_or_else(|| T::zero()) * area * dh * dh);
-
-        Ok(resistance)
-    }
-
-    /// Calculate resistance for laminar flow
-    fn calculate_laminar_resistance(&self, fluid: &Fluid<T>) -> Result<T> {
-        match &self.geometry.cross_section {
-            CrossSection::Rectangular { width, height } => {
-                self.calculate_rectangular_laminar_resistance(fluid, *width, *height)
-            },
-            CrossSection::Circular { diameter } => {
-                self.calculate_circular_laminar_resistance(fluid, *diameter)
-            },
-            _ => {
-                // Use general formula for other shapes
-                self.calculate_stokes_resistance(fluid)
-            }
-        }
-    }
-
-    /// Calculate resistance for rectangular channels (exact solution)
-    fn calculate_rectangular_laminar_resistance(&self, fluid: &Fluid<T>, width: T, height: T) -> Result<T> {
-        // Use actual operating temperature instead of hardcoded 20°C
-        let temperature = T::from_f64(293.15).unwrap_or_else(|| T::zero()); // Default to 20°C if not specified
-        let viscosity = fluid.dynamic_viscosity(temperature)?;
-        let length = self.geometry.length;
-
-        // Exact solution for rectangular channel
-        let aspect_ratio = width / height;
-        let f_re = self.calculate_rectangular_friction_factor(aspect_ratio);
-
-        let area = width * height;
-        let dh = self.geometry.hydraulic_diameter();
-
-        // Correct hydraulic resistance formula: R = (f*Re) * μ * L / (2 * A * Dh^2)
-        let resistance = f_re * viscosity * length / (T::from_f64(2.0).unwrap_or_else(|| T::zero()) * area * dh * dh);
-
-        Ok(resistance)
-    }
-
-    /// Calculate friction factor for rectangular channels
-    fn calculate_rectangular_friction_factor(&self, aspect_ratio: T) -> T {
-        // Proper friction factor calculation using Shah and London (1978) correlation
-        // for laminar flow in rectangular ducts
-        // f*Re = 24*(1 - 1.3553*α + 1.9467*α² - 1.7012*α³ + 0.9564*α⁴ - 0.2537*α⁵)
-        // where α is the aspect ratio (width/height) normalized to [0,1]
-        
-        let alpha = if aspect_ratio >= T::one() { 
-            T::one() / aspect_ratio  // Normalize to [0,1] range
-        } else { 
-            aspect_ratio 
-        };
-        
-        // Shah and London correlation coefficients
-        let c1 = T::from_f64(1.3553).unwrap_or_else(|| T::zero());
-        let c2 = T::from_f64(1.9467).unwrap_or_else(|| T::zero());
-        let c3 = T::from_f64(1.7012).unwrap_or_else(|| T::zero());
-        let c4 = T::from_f64(0.9564).unwrap_or_else(|| T::zero());
-        let c5 = T::from_f64(0.2537).unwrap_or_else(|| T::zero());
-        
-        // Evaluate polynomial using Horner's method for efficiency and numerical stability
-        let polynomial = ((((-c5 * alpha + c4) * alpha - c3) * alpha + c2) * alpha - c1) * alpha + T::one();
-        
-        // f*Re = 24 * polynomial
-        T::from_f64(24.0).unwrap_or_else(|| T::zero()) * polynomial
-    }
-
-    /// Calculate resistance for circular channels (Hagen-Poiseuille)
-    fn calculate_circular_laminar_resistance(&self, fluid: &Fluid<T>, diameter: T) -> Result<T> {
-        // Use actual operating temperature instead of hardcoded 20°C
-        let temperature = T::from_f64(293.15).unwrap_or_else(|| T::zero()); // Default to 20°C if not specified
-        let viscosity = fluid.dynamic_viscosity(temperature)?;
-        let length = self.geometry.length;
-
-        // Hagen-Poiseuille equation: R = (128 * μ * L) / (π * D^4)
-        let pi = T::from_f64(std::f64::consts::PI).unwrap_or_else(|| T::zero());
-        let onehundredtwentyeight = T::from_f64(128.0).unwrap_or_else(|| T::zero());
-        let d4 = num_traits::Float::powf(diameter, T::from_f64(4.0).unwrap_or_else(|| T::zero()));
-
-        let resistance = onehundredtwentyeight * viscosity * length / (pi * d4);
-
-        Ok(resistance)
-    }
-
-    /// Calculate resistance for transitional flow
-    fn calculate_transitional_resistance(&self, fluid: &Fluid<T>) -> Result<T> {
-        // Interpolate between laminar and turbulent
-        let laminar_r = self.calculate_laminar_resistance(fluid)?;
-        let turbulent_r = self.calculate_turbulent_resistance(fluid)?;
-
-        // Standard linear interpolation (could be improved)
-        let blend_factor = T::from_f64(0.5).unwrap_or_else(|| T::zero());
-        Ok(laminar_r * (T::one() - blend_factor) + turbulent_r * blend_factor)
-    }
-
-    /// Calculate resistance for turbulent flow using Darcy-Weisbach equation
-    fn calculate_turbulent_resistance(&self, fluid: &Fluid<T>) -> Result<T> {
-        let reynolds = self.flow_state.reynolds_number.ok_or_else(|| {
-            cfd_core::error::Error::InvalidConfiguration("Reynolds number required for turbulent flow".to_string())
-        })?;
-
-        // Use Darcy-Weisbach equation with friction factor correlation
-        let area = self.geometry.area();
-        let dh = self.geometry.hydraulic_diameter();
-        let length = self.geometry.length;
-        let density = fluid.density;
-
-        // Calculate friction factor using Swamee-Jain approximation for smooth pipes
-        let relative_roughness = self.geometry.surface.roughness / dh;
-        let friction_factor = self.calculate_turbulent_friction_factor(reynolds, relative_roughness);
-
-        // Darcy-Weisbach resistance: R = f * L * ρ / (2 * A * Dh^2)
-        let resistance = friction_factor * length * density /
-                        (T::from_f64(2.0).unwrap_or_else(|| T::zero()) * area * dh * dh);
-
-        Ok(resistance)
-    }
-
-    /// Calculate turbulent friction factor using Swamee-Jain approximation
-    fn calculate_turbulent_friction_factor(&self, reynolds: T, relative_roughness: T) -> T {
-        // Swamee-Jain approximation to Colebrook-White equation
-        // Valid for: 5000 < Re < 10^8, 10^-6 < ε/D < 10^-2
-        let term1 = relative_roughness / T::from_f64(3.7).unwrap_or_else(|| T::zero());
-        let term2 = T::from_f64(5.74).unwrap_or_else(|| T::zero()) / num_traits::Float::powf(reynolds, T::from_f64(0.9).unwrap_or_else(|| T::zero()));
-        let log_term = num_traits::Float::ln(term1 + term2);
-        T::from_f64(0.25).unwrap_or_else(|| T::zero()) / num_traits::Float::powf(log_term, T::from_f64(2.0).unwrap_or_else(|| T::zero()))
-    }
-
-    /// Calculate resistance for slip flow (rarefied gas) using Knudsen number corrections
-    fn calculate_slip_flow_resistance(&self, fluid: &Fluid<T>) -> Result<T> {
-        // Calculate Knudsen number: Kn = λ / Dh
-        // where λ is the mean free path
-        let dh = self.geometry.hydraulic_diameter();
-        let _temperature = T::from_f64(293.15).unwrap_or_else(|| T::zero()); // Default temperature (for future use)
-
-        // Estimate mean free path for air at standard conditions
-        // λ ≈ μ * sqrt(π * R * T / (2 * M)) / P
-        // Simplified approximation: λ ≈ 68 nm at STP
-        let mean_free_path = T::from_f64(68e-9).unwrap_or_else(|| T::zero()); // meters
-        let knudsen = mean_free_path / dh;
-
-        // Get laminar resistance as base
-        let laminar_r = self.calculate_laminar_resistance(fluid)?;
-
-        // Apply slip flow correction based on Knudsen number
-        // For 0.01 < Kn < 0.1 (slip flow regime)
-        let slip_correction = if knudsen > T::from_f64(0.01).unwrap_or_else(|| T::zero()) && knudsen < T::from_f64(0.1).unwrap_or_else(|| T::zero()) {
-            // Beskok-Karniadakis model: f_app = f_continuum * (1 + α * Kn)^(-1)
-            let alpha = T::from_f64(1.358).unwrap_or_else(|| T::zero()); // Slip coefficient
-            T::one() / (T::one() + alpha * knudsen)
-        } else if knudsen >= T::from_f64(0.1).unwrap_or_else(|| T::zero()) {
-            // Transition to free molecular flow
-            T::from_f64(0.5).unwrap_or_else(|| T::zero()) // Significant reduction
-        } else {
-            // Continuum flow
-            T::one()
-        };
-
-        Ok(laminar_r * slip_correction)
-    }
-
-    /// Get shape factor for different cross-sections
-    fn get_shape_factor(&self) -> T {
-        match &self.geometry.cross_section {
-            CrossSection::Rectangular { .. } => T::from_f64(24.0).unwrap_or_else(|| T::zero()),
-            CrossSection::Circular { .. } => T::from_f64(16.0).unwrap_or_else(|| T::zero()),
-            CrossSection::Elliptical { .. } => T::from_f64(20.0).unwrap_or_else(|| T::zero()),
-            CrossSection::Trapezoidal { .. } => T::from_f64(22.0).unwrap_or_else(|| T::zero()),
-            CrossSection::Custom { .. } => T::from_f64(20.0).unwrap_or_else(|| T::zero()),
-        }
-    }
-
-    /// Set Reynolds number for flow state
-    pub fn set_reynolds_number(&mut self, reynolds: T) {
-        self.flow_state.reynolds_number = Some(reynolds);
-        self.flow_state.flow_regime = self.classify_flow_regime(reynolds);
-    }
-
-    /// Get current flow regime
-    pub fn flow_regime(&self) -> &FlowRegime {
-        &self.flow_state.flow_regime
-    }
-
-    /// Enable entrance effects modeling
-    pub fn enable_entrance_effects(&mut self) {
-        self.numerical_params.entrance_effects = true;
-        self.flow_state.entrance_effects = true;
-    }
-
-    /// Enable surface tension effects
-    pub fn enable_surface_tension_effects(&mut self) {
-        self.numerical_params.surface_tension_effects = true;
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use approx::assert_relative_eq;
-
-    #[test]
-    fn test_rectangular_geometry() {
-        let geometry = ChannelGeometry::rectangular(0.001, 100e-6, 50e-6, 1e-6);
-
-        assert_relative_eq!(geometry.area(), 5e-9, epsilon = 1e-15);
-        assert_relative_eq!(geometry.hydraulic_diameter(), 200e-6 / 3.0, epsilon = 1e-15);
-        assert_relative_eq!(geometry.wetted_perimeter(), 300e-6, epsilon = 1e-15);
-    }
-
-    #[test]
-    fn test_circular_geometry() {
-        let geometry = ChannelGeometry::circular(0.001, 100e-6, 1e-6);
-
-        let expected_area = std::f64::consts::PI * (50e-6_f64).powi(2);
-        assert_relative_eq!(geometry.area(), expected_area, epsilon = 1e-15);
-        assert_relative_eq!(geometry.hydraulic_diameter(), 100e-6, epsilon = 1e-15);
-
-        let expected_perimeter = std::f64::consts::PI * 100e-6;
-        assert_relative_eq!(geometry.wetted_perimeter(), expected_perimeter, epsilon = 1e-15);
-    }
-
-    #[test]
-    fn test_channel_creation() {
-        let geometry = ChannelGeometry::rectangular(0.001, 100e-6, 50e-6, 1e-6);
-        let channel = Channel::new(geometry);
-
-        assert_eq!(channel.flow_state.flow_regime, FlowRegime::Laminar);
-        assert!(!channel.flow_state.entrance_effects);
-        assert_eq!(channel.numerical_params.discretization_points, 100);
-    }
-
-    #[test]
-    fn test_flow_regime_classification() {
-        let geometry = ChannelGeometry::rectangular(0.001, 100e-6, 50e-6, 1e-6);
-        let channel = Channel::new(geometry);
-
-        assert_eq!(channel.classify_flow_regime(0.05), FlowRegime::Stokes);
-        assert_eq!(channel.classify_flow_regime(100.0), FlowRegime::Laminar);
-        assert_eq!(channel.classify_flow_regime(3000.0), FlowRegime::Transitional);
-        assert_eq!(channel.classify_flow_regime(5000.0), FlowRegime::Turbulent);
-    }
-
-    #[test]
-    fn test_rectangular_friction_factor() {
-        let geometry = ChannelGeometry::rectangular(0.001, 100e-6, 50e-6, 1e-6);
-        let channel = Channel::new(geometry);
-
-        // Test square channel (aspect ratio = 1)
-        let f_re_square = channel.calculate_rectangular_friction_factor(1.0);
-        assert!(f_re_square > 0.0);
-
-        // Test wide channel (aspect ratio = 2)
-        let f_re_wide = channel.calculate_rectangular_friction_factor(2.0);
-        assert!(f_re_wide > 0.0);
-        // Note: Simplified model may not preserve exact ordering, just check positivity
-    }
-
-    #[test]
-    fn test_circular_resistance_calculation() {
-        let geometry = ChannelGeometry::circular(0.001, 100e-6, 1e-6);
-        let channel = Channel::<f64>::new(geometry);
-        let fluid = cfd_core::fluid::Fluid::<f64>::water().expect("Failed to create water fluid");
-
-        let resistance = channel.calculate_circular_laminar_resistance(&fluid, 100e-6).expect("Failed to complete operation");
-
-        // Should be positive and reasonable for water flow
-        assert!(resistance > 0.0);
-        assert!(resistance < 1e15); // Reasonable upper bound
-    }
-
-    #[test]
-    fn test_rectangular_resistance_calculation() {
-        let geometry = ChannelGeometry::rectangular(0.001, 100e-6, 50e-6, 1e-6);
-        let channel = Channel::<f64>::new(geometry);
-        let fluid = cfd_core::fluid::Fluid::<f64>::water().expect("Failed to create water fluid");
-
-        let resistance = channel.calculate_rectangular_laminar_resistance(&fluid, 100e-6, 50e-6).expect("Failed to complete operation");
-
-        // Should be positive and reasonable for water flow
-        assert!(resistance > 0.0);
-        assert!(resistance < 1e15); // Reasonable upper bound
-    }
-
-    #[test]
-    fn test_resistance_calculation_different_regimes() {
-        let geometry = ChannelGeometry::rectangular(0.001, 100e-6, 50e-6, 1e-6);
-        let mut channel = Channel::<f64>::new(geometry);
-        let fluid = cfd_core::fluid::Fluid::<f64>::water().expect("Failed to create water fluid");
-
-        // Test different flow regimes
-        channel.set_reynolds_number(0.1); // Stokes
-        let stokes_r = channel.calculate_resistance(&fluid).expect("Failed to complete operation");
-
-        channel.set_reynolds_number(100.0); // Laminar
-        let laminar_r = channel.calculate_resistance(&fluid).expect("Failed to complete operation");
-
-        channel.set_reynolds_number(3000.0); // Transitional
-        let transitional_r = channel.calculate_resistance(&fluid).expect("Failed to complete operation");
-
-        channel.set_reynolds_number(5000.0); // Turbulent
-        let turbulent_r = channel.calculate_resistance(&fluid).expect("Failed to complete operation");
-
-        // All should be positive
-        assert!(stokes_r > 0.0);
-        assert!(laminar_r > 0.0);
-        assert!(transitional_r > 0.0);
-        assert!(turbulent_r > 0.0);
-
-        // Turbulent should generally have higher resistance than laminar
-        assert!(turbulent_r >= laminar_r);
-    }
-
-    #[test]
-    fn test_shape_factors() {
-        let rect_geometry = ChannelGeometry::rectangular(0.001, 100e-6, 50e-6, 1e-6);
-        let rect_channel = Channel::new(rect_geometry);
-
-        let circ_geometry = ChannelGeometry::circular(0.001, 100e-6, 1e-6);
-        let circ_channel = Channel::new(circ_geometry);
-
-        let rect_factor = rect_channel.get_shape_factor();
-        let circ_factor = circ_channel.get_shape_factor();
-
-        assert_relative_eq!(rect_factor, 24.0, epsilon = 1e-10);
-        assert_relative_eq!(circ_factor, 16.0, epsilon = 1e-10);
-    }
-
-    #[test]
-    fn test_entrance_effects() {
-        let geometry = ChannelGeometry::rectangular(0.001, 100e-6, 50e-6, 1e-6);
-        let mut channel = Channel::new(geometry);
-
-        // Initially disabled
-        assert!(!channel.flow_state.entrance_effects);
-
-        // Enable entrance effects
-        channel.enable_entrance_effects();
-        assert!(channel.flow_state.entrance_effects);
-        assert!(channel.numerical_params.entrance_effects);
-    }
-
-    #[test]
-    fn test_surface_tension_effects() {
-        let geometry = ChannelGeometry::rectangular(0.001, 100e-6, 50e-6, 1e-6);
-        let mut channel = Channel::new(geometry);
-
-        // Initially disabled
-        assert!(!channel.numerical_params.surface_tension_effects);
-
-        // Enable surface tension effects
-        channel.enable_surface_tension_effects();
-        assert!(channel.numerical_params.surface_tension_effects);
-    }
-
-    #[test]
-    fn test_elliptical_geometry() {
-        let geometry = ChannelGeometry {
-            channel_type: ChannelType::Straight,
-            length: 0.001,
-            cross_section: CrossSection::Elliptical {
-                major_axis: 200e-6,
-                minor_axis: 100e-6
-            },
-            surface: SurfaceProperties {
-                roughness: 1e-6,
-                contact_angle: None,
-                surface_energy: None,
-                wettability: Wettability::Hydrophilic,
-            },
-            variations: Vec::new(),
-        };
-
-        let expected_area = std::f64::consts::PI * 200e-6 * 100e-6 / 4.0;
-        assert_relative_eq!(geometry.area(), expected_area, epsilon = 1e-15);
-
-        // Hydraulic diameter should be reasonable
-        let dh = geometry.hydraulic_diameter();
-        assert!(dh > 0.0);
-        assert!(dh < 300e-6); // Should be reasonable for ellipse
-    }
-
-    #[test]
-    fn test_trapezoidal_geometry() {
-        let geometry = ChannelGeometry {
-            channel_type: ChannelType::Straight,
-            length: 0.001,
-            cross_section: CrossSection::Trapezoidal {
-                top_width: 150e-6,
-                bottom_width: 50e-6,
-                height: 100e-6,
-            },
-            surface: SurfaceProperties {
-                roughness: 1e-6,
-                contact_angle: None,
-                surface_energy: None,
-                wettability: Wettability::Hydrophilic,
-            },
-            variations: Vec::new(),
-        };
-
-        let expected_area = (150e-6 + 50e-6) * 100e-6 / 2.0;
-        assert_relative_eq!(geometry.area(), expected_area, epsilon = 1e-15);
-
-        // Should have reasonable hydraulic diameter and perimeter
-        assert!(geometry.hydraulic_diameter() > 0.0);
-        assert!(geometry.wetted_perimeter() > 0.0);
-    }
-}
