+++ crates/cfd-3d/src/fem/element.rs
@@ -68,11 +68,56 @@ impl<T: RealField + FromPrimitive + Copy> FluidElement<T> {
         self.volume
     }
     
-    /// Calculate shape function derivatives
+    /// Calculate shape function derivatives for tetrahedral element
+    /// Based on Zienkiewicz & Taylor, "The Finite Element Method", Vol. 1, 6th Ed.
     pub fn calculate_shape_derivatives(&mut self, vertices: &[Vector3<T>]) {
-        // For linear tetrahedron, derivatives are constant
-        // This is a placeholder - full implementation needed
-        self.dN_dx = Matrix3::identity();
+        if vertices.len() != 4 {
+            // For non-tetrahedral elements, use identity as fallback
+            self.dN_dx = Matrix3::identity();
+            return;
+        }
+        
+        // For linear tetrahedron with vertices at (x1,y1,z1), ..., (x4,y4,z4)
+        // Shape functions: N_i = (a_i + b_i*x + c_i*y + d_i*z) / (6*V)
+        // where V is the element volume
+        
+        let v1 = vertices[0];
+        let v2 = vertices[1];
+        let v3 = vertices[2];
+        let v4 = vertices[3];
+        
+        // Calculate volume using scalar triple product
+        let volume = ((v2 - v1).cross(&(v3 - v1))).dot(&(v4 - v1)) / T::from_f64(6.0).unwrap_or_else(T::one);
+        
+        if volume.abs() < T::from_f64(1e-12).unwrap_or_else(T::zero) {
+            // Degenerate element
+            self.dN_dx = Matrix3::zeros();
+            return;
+        }
+        
+        // Calculate shape function derivatives (constant for linear elements)
+        // dN1/dx = (y2(z3-z4) + y3(z4-z2) + y4(z2-z3)) / (6*V)
+        // Similar for other derivatives
+        let six_v = T::from_f64(6.0).unwrap_or_else(T::one) * volume;
+        
+        // Store as columns of the derivative matrix
+        self.dN_dx = Matrix3::from_columns(&[
+            Vector3::new(
+                (v2.y * (v3.z - v4.z) + v3.y * (v4.z - v2.z) + v4.y * (v2.z - v3.z)) / six_v,
+                (v2.x * (v4.z - v3.z) + v3.x * (v2.z - v4.z) + v4.x * (v3.z - v2.z)) / six_v,
+                (v2.x * (v3.y - v4.y) + v3.x * (v4.y - v2.y) + v4.x * (v2.y - v3.y)) / six_v
+            ),
+            Vector3::new(
+                (v1.y * (v4.z - v3.z) + v3.y * (v1.z - v4.z) + v4.y * (v3.z - v1.z)) / six_v,
+                (v1.x * (v3.z - v4.z) + v3.x * (v4.z - v1.z) + v4.x * (v1.z - v3.z)) / six_v,
+                (v1.x * (v4.y - v3.y) + v3.x * (v1.y - v4.y) + v4.x * (v3.y - v1.y)) / six_v
+            ),
+            Vector3::new(
+                (v1.y * (v2.z - v4.z) + v2.y * (v4.z - v1.z) + v4.y * (v1.z - v2.z)) / six_v,
+                (v1.x * (v4.z - v2.z) + v2.x * (v1.z - v4.z) + v4.x * (v2.z - v1.z)) / six_v,
+                (v1.x * (v2.y - v4.y) + v2.x * (v4.y - v1.y) + v4.x * (v1.y - v2.y)) / six_v
+            )
+        ]);
     }
     
     /// Calculate strain rate from velocity gradient
@@ -81,13 +126,43 @@ impl<T: RealField + FromPrimitive + Copy> FluidElement<T> {
         (velocity_gradient + velocity_gradient.transpose()) * half
     }
     
-    /// Calculate element stiffness matrix contribution
+    /// Calculate element stiffness matrix contribution for Stokes flow
+    /// Based on Hughes, "The Finite Element Method", Dover Publications, 2000
     pub fn stiffness_contribution(&self, viscosity: T) -> DMatrix<T> {
         let n_dof = self.nodes.len() * constants::VELOCITY_COMPONENTS;
-        let k_e = DMatrix::zeros(n_dof, n_dof);
+        let mut k_e = DMatrix::zeros(n_dof, n_dof);
         
-        // Placeholder implementation - needs proper Galerkin formulation
-        // K_e = ∫_Ωe B^T D B dΩ
+        // For Stokes flow: K_e = ∫_Ωe B^T D B dΩ
+        // where B is the strain-displacement matrix and D is the material matrix
+        
+        // For linear tetrahedral elements, we use single-point integration
+        // This is exact for linear elements
+        
+        if self.nodes.len() == 4 {
+            // Tetrahedral element
+            let two = T::from_f64(2.0).unwrap_or_else(T::one);
+            let factor = viscosity * self.volume;
+            
+            // Build B matrix (strain-displacement matrix)
+            // For 3D: B relates nodal velocities to strain rates
+            // Size: 6 x (4*3) for tetrahedral element
+            
+            // Simplified implementation for viscous term only
+            // Full implementation would include pressure terms
+            for i in 0..4 {
+                for j in 0..4 {
+                    for d in 0..3 {
+                        let row = i * 3 + d;
+                        let col = j * 3 + d;
+                        
+                        // Viscous contribution: μ ∫ ∇Ni · ∇Nj dΩ
+                        // For constant shape function derivatives in tetrahedral elements
+                        let grad_prod = self.dN_dx[(d, i)] * self.dN_dx[(d, j)];
+                        k_e[(row, col)] += factor * grad_prod;
+                    }
+                }
+            }
+        }
         
         k_e
     }
